Split oversized modules into focused submodules (#145)
- mahjong-server round: split 1850-line mod.rs into turn (draw/discard/
  kan/pei), calls (call resolution), win (riichi/tsumo), and draws
  (exhaustive/special draws)
- mahjong-client renderer: split 2185-line mod.rs into board, tiles,
  result, setup, and tests submodules
- mahjong-client game: converted 2046-line game.rs into a module
  directory with events, input, labels, setup, and tests

Pure code movement; no behavior changes. Cross-module private methods
are now pub(super), scoped to their parent module.
